Prepared statements, working apostrophe test.
All CQL queries are now prepared statements. This not only
makes the code look a million times better, but it also fixes
two bugs.

1) We couldn't use certain punctuation in titles or text.
2) We were open to brutal CQL injection attacks.

+++ src/main/java/WordCount.java
@@ -72,7 +72,7 @@ public class WordCount {
     private void countWordForTitle(String word, String title) {
         long currentCount = persistence.getCurrentCountForWord(word, title);
         persistence.incrementCountForWord(word, title);
-        persistence.updateTopTen(word, currentCount + 1L, title);
+        persistence.updateTopTen(word, (int)currentCount + 1, title);
     }
 
     private void countWordForGrandTotal(String word) {
